Rename ServiceContext to ContextProvider
