+++ web/src/index.js
@@ -7,7 +7,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import MuiPickersUtilsProvider from 'material-ui-pickers/utils/MuiPickersUtilsProvider';
 import MomentUtils from 'material-ui-pickers/utils/moment-utils';
-import ContextProvider from './shared/ServiceContext';
+import ContextProvider from './shared/ContextProvider';
 
 import './index.css';
 import App from './app/App';
