+++ web/src/scans/Scanner.js
@@ -5,7 +5,7 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { withContext } from '../shared/ServiceContext';
+import { withContext } from '../shared/ContextProvider';
 import moment from 'moment';
 
 import { withStyles } from '@material-ui/core/styles';
