+++ web/src/veterans/VeteranDialog.js
@@ -21,7 +21,7 @@ import {
 } from '@material-ui/core';
 
 import { validateEmail, validatePhoneNumber, userCanView } from '../util';
-import { withContext } from '../shared/ServiceContext';
+import { withContext } from '../shared/ContextProvider';
 
 import CircularProgress from '@material-ui/core/CircularProgress';
 import ConfirmDialog from '../shared/ConfirmDialog';
